strict convert distance from int to float

+++ realsense2_camera/src/base_realsense_node.cpp
@@ -2568,7 +2568,7 @@ void BaseRealSenseNode::publishAIdata(rs2::frame f, const ros::Time& header_time
                         msg_qr.y = ai_data.y/1000;
                         msg_qr.z = ai_data.z/1000;
                         msg_qr.theta = ai_data.degree;
-                        msg_qr.distance = ai_data.distance/1000;
+                        msg_qr.distance = static_cast<float>(ai_data.distance)/1000;
                         msg_qr_list.list.push_back(msg_qr);
                         ai_box_info_start += sizeof(ai_data);
                     }
